# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "Task API with Supabase-backed authentication. Use the Authorize button with a Bearer access token (returned by /auth/signin) to test protected routes."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the access_token returned from /auth/signin here (without the word 'Bearer')."
      }
    },
    "schemas": {
      "SignUpRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "test@example.com" },
          "password": { "type": "string", "format": "password", "example": "SuperSecret123!" }
        }
      },
      "SignInRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "test@example.com" },
          "password": { "type": "string", "format": "password", "example": "SuperSecret123!" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "AuthSuccessResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/": {
      "get": {
        "summary": "API info",
        "tags": ["General"],
        "responses": {
          "200": {
            "description": "Basic API metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "version": { "type": "string" },
                    "endpoints": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "tags": ["General"],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string", "example": "ok" } }
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Create a new account",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SignUpRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthSuccessResponse" } }
            }
          },
          "400": {
            "description": "Sign up failed (e.g. invalid email, weak password, user already exists)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "summary": "Sign in and receive an access token",
        "description": "Copy the returned access_token and paste it into the Authorize (padlock) button to call protected routes.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SignInRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in successfully",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthSuccessResponse" } }
            }
          },
          "400": {
            "description": "Missing email or password",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "401": {
            "description": "Invalid login credentials",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/auth/signout": {
      "post": {
        "summary": "Sign out the current session",
        "description": "Requires a valid Bearer access token. Click Authorize and paste the token from /auth/signin first.",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Signed out successfully",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } }
            }
          },
          "400": {
            "description": "Sign out failed",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired token",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public info (no auth required)",
        "tags": ["Public"],
        "responses": {
          "200": {
            "description": "Publicly accessible message",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } }
            }
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get the authenticated user's profile",
        "description": "Requires a valid Bearer access token. Click Authorize and paste the token from /auth/signin first.",
        "tags": ["Protected"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Authenticated user info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired token",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    }
  }
}